feat: atualização flyway

+++ src/main/resources/db/migration/V04__criar_e_registrar_usarios.sql
@@ -1,6 +1,6 @@
 -- Criação da tabela 'usuario'
 CREATE TABLE usuario (
-  codigo BIGINT PRIMARY KEY,
+  codigo SERIAL PRIMARY KEY,
   nome VARCHAR(50) NOT NULL,
   email VARCHAR(50) NOT NULL,
   senha VARCHAR(150) NOT NULL
